requirements.txt
anthropic>=0.39.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
gitpython>=3.1.0
docker>=7.0.0
